refactor : Refactor codes (13) and add InputProcessor
