+++ src/main/java/racingcar/view/InputView.java
@@ -2,11 +2,9 @@ package racingcar.view;
 
 import java.util.Arrays;
 import java.util.List;
-import java.util.Scanner;
 
 public class InputView {
     private static final String DELIMITER_FOR_INPUT_CAR = ",";
-    private static final Scanner scanner = new Scanner(System.in);
 
     private List<String> carNames;
     private int carCount;
@@ -15,16 +13,6 @@ public class InputView {
     public InputView(String carNames, String tryCounts) {
         insertCarNames(carNames);
         insertTryCount(tryCounts);
-    }
-
-    public static String getCarNamesFromUser() {
-        System.out.println("자동차 이름을 입력하세요, (단, 자동차 이름은 쉼표(,)로 구분해야 합니다.)");
-        return scanner.nextLine();
-    }
-
-    public static String getTryCountsFromUser() {
-        System.out.println("몇 회를 시도하실 건가요?");
-        return scanner.nextLine();
     }
 
     private void insertCarNames(String carNameString) {
